first

+++ src/main/java/com/cico/controller/TaskController.java
@@ -57,12 +57,11 @@ public class TaskController {
 	}
 
 	@PostMapping("/studentTaskSubmittion")
-	public ResponseEntity<?> StudentTaskSubmittion(@RequestParam("taskId") Integer taskId,
+	public ResponseEntity<?> StudentTaskSubmittion(@RequestParam("taskId") Long taskId,
 			@RequestParam("studentId") Integer studentId,
 			@RequestParam(name = "submittionFileName", required = false) MultipartFile file,
 			@RequestParam("taskDescription") String taskDescription) {
-		return taskService.studentTaskSubmittion(taskId,studentId, file, taskDescription);
-
+		return taskService.studentTaskSubmittion(taskId, studentId, file, taskDescription);
 
 	}
 
@@ -85,27 +84,25 @@ public class TaskController {
 			@RequestParam("taskId") Integer taskId) {
 		return taskService.deleteTaskQuestion(taskId, questionId);
 	}
-	
+
 	@GetMapping("/getSubmitedTaskForStudent")
-	public ResponseEntity<?> getSubmitedTaskForStudent(@RequestParam("studentId") Integer studentId){
-		return taskService.getSubmitedTaskForStudent(studentId);	
+	public ResponseEntity<?> getSubmitedTaskForStudent(@RequestParam("studentId") Integer studentId) {
+		return taskService.getSubmitedTaskForStudent(studentId);
 	}
-	
+
 	@GetMapping("/getAllSubmitedTask")
 	public ResponseEntity<?> getAllSubmitedTasks() {
 		return taskService.getAllSubmitedTasks();
 	}
 
-
 	@GetMapping("/getAllSubmissionTaskStatus")
-	public  ResponseEntity<?>getAllSubmissionTaskStatus(){
-		 return  taskService.getAllSubmissionTaskStatus();
+	public ResponseEntity<?> getAllSubmissionTaskStatus() {
+		return taskService.getAllSubmissionTaskStatus();
 	}
 
-		
 	@PutMapping("/updateSubmitedAssignmentStatus")
 	public ResponseEntity<?> updateSubmitedAssignmentStatus(@RequestParam("submissionId") Integer submissionId,
-			@RequestParam("status") String status,@RequestParam("review") String review){
-		return taskService.updateSubmitedTaskStatus(submissionId,status,review);
+			@RequestParam("status") String status, @RequestParam("review") String review) {
+		return taskService.updateSubmitedTaskStatus(submissionId, status, review);
 	}
 }
